% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpeak.R
\name{qpeak}
\alias{qpeak}
\title{Peak into contribution of any variable to a predictive model}
\usage{
qpeak(model, X, feature, which.class = 1L, predict, ...)
}
\arguments{
\item{model}{Trained predictive model.}

\item{X}{data.frame which was used to train the predictive model. It should not
contain the column which is predicted by the model.}

\item{feature}{Name of the feature of interest.}

\item{which.class}{Integer specifying which column of the matrix of predicted
probabilities to use as the "focus" class. Default is to use
    the first class. Only used for classification problems.}

\item{predict}{Custom predict function that obtaines class probabilities from model}

\item{...}{Optional parameters to the cutsom predict function.}
}
\description{
Peak into contribution of any variable to a predictive model
}
