% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpeak_shiny.R
\name{qpeak_shiny}
\alias{qpeak_shiny}
\title{Shiny interface to quickpeak}
\usage{
qpeak_shiny(model, X, FUN, ...)
}
\arguments{
\item{model}{Trained predictive model.}

\item{X}{data.frame which was used to train the predictive model. It should not
contain the column which is predicted by the model.}

\item{FUN}{Custom predict function that obtaines class probabilities from model}

\item{...}{Optional parameters to the}
}
\description{
Shiny interface to quickpeak
}
