% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpeak.R
\name{replace_continuous}
\alias{replace_continuous}
\title{Helper functions
Replaces continuous variables with their mediam}
\usage{
replace_continuous(x)
}
\description{
Helper functions
Replaces continuous variables with their mediam
}
