% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tree.R
\name{show_tree}
\alias{show_tree}
\title{Visual representation of a single tree from a random forest model}
\usage{
show_tree(model, tree)
}
\arguments{
\item{model}{random forest model as produced by ranger package}

\item{tree}{integer reflecting which tree to visualise}
}
\description{
Visual representation of a single tree from a random forest model
}
